Hardcoded Effect and Light Selector
So this is a really bad commit with too much stuff.... Hardcoded in a lighting effect for the time being (adding an editor is on the feature list). Added a light selector so you can choose which lights respond to the alerts. Did some more cleanup of old code as well.